+++ lib/redix/utils.ex
@@ -53,14 +53,14 @@ defmodule Redix.Utils do
     {redix_opts, connection_opts}
   end
 
-  @spec connect(Keyword.t) :: {:ok, :gen_tcp.socket} | {:error, term} | {:stop, term, %{}}
+  @spec connect(Keyword.t) :: {:ok, :ssl.sslsocket} | {:error, term} | {:stop, term, %{}}
   def connect(opts) do
     host = opts |> Keyword.fetch!(:host) |> String.to_charlist()
     port = Keyword.fetch!(opts, :port)
     socket_opts = @socket_opts ++ Keyword.fetch!(opts, :socket_opts)
     timeout = opts[:timeout] || @default_timeout
 
-    with {:ok, socket} <-:gen_tcp.connect(host, port, socket_opts, timeout),
+    with {:ok, socket} <-:ssl.connect(host, port, socket_opts, timeout),
          :ok <- setup_socket_buffers(socket) do
       result =
         with :ok <- if(opts[:password], do: auth(socket, opts[:password]), else: :ok),
@@ -81,9 +81,8 @@ defmodule Redix.Utils do
 
   # Setups the `:buffer` option of the given socket.
   defp setup_socket_buffers(socket) do
-    with {:ok, opts} <- :inet.getopts(socket, [:sndbuf, :recbuf, :buffer]) do
-      [sndbuf: sndbuf, recbuf: recbuf, buffer: buffer] = opts
-      :inet.setopts(socket, buffer: buffer |> max(sndbuf) |> max(recbuf))
+    with {:ok, opts} <- :ssl.getopts(socket, [:sndbuf, :recbuf, :buffer]) do
+      :ssl.setopts(socket, buffer: opts[:buffer] |> max(opts[:sndbuf]) |> max(opts[:recbuf]))
     end
   end
 
@@ -106,12 +105,12 @@ defmodule Redix.Utils do
   end
 
   defp auth(socket, password) do
-    with :ok <- :gen_tcp.send(socket, Redix.Protocol.pack(["AUTH", password])),
+    with :ok <- :ssl.send(socket, Redix.Protocol.pack(["AUTH", password])),
          do: recv_ok_response(socket)
   end
 
   defp select(socket, database) do
-    with :ok <- :gen_tcp.send(socket, Redix.Protocol.pack(["SELECT", database])),
+    with :ok <- :ssl.send(socket, Redix.Protocol.pack(["SELECT", database])),
          do: recv_ok_response(socket)
   end
 
@@ -120,7 +119,7 @@ defmodule Redix.Utils do
   end
 
   defp recv_ok_response(socket, continuation) do
-    with {:ok, data} <- :gen_tcp.recv(socket, 0) do
+    with {:ok, data} <- :ssl.recv(socket, 0) do
       parser = continuation || &Redix.Protocol.parse/1
       case parser.(data) do
         {:ok, "OK", ""} ->